Don't start cowboy as part of piqi-rpc supervision tree
This fixes all kinds of nasty race conditions in stopping piqi-rpc.

Additionally, see [this](https://github.com/extend/cowboy/pull/265) issue on
why this is not supported and not the cowboy way of doing things.

Also added tests and refactored the `piqi_rpc_http:get_http_env` to not require
to always pass the default explicitly.