Add function field serialisation / default helpers
This adds helper functions for (de)serialisations and default
generation of function input/output/error fields to `$MOD_piqi_rpc.erl`.

The rationale is that, unlike types, the actual function (in particular, its
name) is something which exists only in `$MOD_piqi_rpc.erl`, so it makes sense to
provide these helpers and abstract from the actual type definitions the
function fields.

This also allows to slightly simplify some of the other generated code.